pyproject.toml
[tool.poetry]
name = "demo-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
sqlalchemy = "^2.0.31"
pymysql = "^1.1.1"
pydantic = {extras = ["email"], version = "^2.9.2"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.12"
aiomysql = "^0.2.0"
requests = "^2.32.3"
google-auth = "^2.36.0"
httpx = "^0.27.2"
langserve = "^0.3.0"
langchain = "^0.3.7"
langchain-openai = "^0.2.9"
google = "^3.0.0"
langchain-community = "^0.3.7"
duckduckgo-search = "^6.3.5"
langchain-teddynote = "^0.3.30"
faiss-cpu = "^1.9.0.post1"
sse-starlette = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
